Laboratorios/Lab-Sistema-Inscripcion/service-notification/requirements.txt
pika==1.3.1
Flask==2.0.1
python-dotenv==0.19.0
# smtplib
# email.mime.text
# email.mime.multipart